+++ arc1pyqt/ProgPanels/Retention.py
@@ -297,7 +297,7 @@ class Retention(BaseProgPanel):
         resLayout.addItem(statsLayout)
 
         resultWindow = QtWidgets.QWidget()
-        resultWindow.setGeometry(100,100,1000*APP.scalingFactor, 400)
+        resultWindow.setGeometry(100,100,int(1000*APP.scalingFactor), 400)
         resultWindow.setWindowTitle("Retention: W="+ str(w) + " | B=" + str(b))
         resultWindow.setWindowIcon(Graphics.getIcon('appicon'))
         resultWindow.show()
